requirements.txt
Flask==3.0.0
python-telegram-bot==13.15
gunicorn==21.2.0
requests==2.31.0
